+++ src/components/index.js
@@ -1,1 +1,5 @@
 export { default as PageLayout } from './layout';
+export { default as AdminNavbar } from './navbar';
+export { default as AdminSidebar } from './sidebar';
+export { default as OrdersDashboard } from './orders';
+export { default as ProductsDashboard } from './products';
